+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.90 (Pending):
+* Add method `AbstractServerTest.eventStub` that retrieves the test event stub.
 * Add method to AbstractServerTest and AbstractSharedTest to retrieve objects by type literal.
 
 ## 0.89: